refactor(ds-global): add groups tier, move KeyboardKey(s) to real tiers
KeyboardKey is an ontology Component and KeyboardKeys an ontology Group
(ds:global.group.keyboard_keys). Move KeyboardKey to component/ and
KeyboardKeys to a new group/ tier, and add the 'groups' key to the shared
storybook-config sidebar order (after components, before patterns) so
Group-type stories sort into their own section. KeyboardKeys is a Group
only, not a component. Cross-tier import uses #lib/.
