+++ packages/react/ds-global/src/lib/component/KeyboardKey/constants.ts
