+++ packages/react/ds-global/src/lib/group/KeyboardKeys/KeyboardKeys.stories.tsx
@@ -1,9 +1,9 @@
 import type { Meta, StoryFn } from "@storybook/react-vite";
-import { KeyboardKey } from "../KeyboardKey/index.js";
+import { KeyboardKey } from "#lib/component/KeyboardKey/index.js";
 import Component from "./KeyboardKeys.js";
 
 const meta = {
-  title: "_work_in_progress/KeyboardKeys",
+  title: "groups/KeyboardKeys",
   component: Component,
   tags: ["autodocs"],
 } satisfies Meta<typeof Component>;
