+++ packages/react/ds-global/src/lib/index.ts
@@ -3,8 +3,6 @@ export * from "./_work_in_progress/ChatSection/index.js";
 export * from "./_work_in_progress/DescriptionSection/index.js";
 export * from "./_work_in_progress/grid/index.js";
 export * from "./_work_in_progress/IconSection/index.js";
-export * from "./_work_in_progress/KeyboardKey/index.js";
-export * from "./_work_in_progress/KeyboardKeys/index.js";
 export * from "./_work_in_progress/Label/index.js";
 export * from "./_work_in_progress/Link/index.js";
 export * from "./_work_in_progress/Rule/index.js";
@@ -19,7 +17,9 @@ export * from "./component/Button/index.js";
 export * from "./component/Card/index.js";
 export * from "./component/Chip/index.js";
 export * from "./component/Icon/index.js";
+export * from "./component/KeyboardKey/index.js";
 export * from "./component/Tile/index.js";
 export * from "./component/Tooltip/index.js";
+export * from "./group/KeyboardKeys/index.js";
 export * from "./pattern/Timeline/index.js";
 export type * from "./types/index.js";
